Looks like Merge is proto done and  foundation for Split has been layed.  Oligos included.
Kept with the concept of creating shallowCopies of Oligos, in addition to Strands, on Merge.  It makes book keeping old _strand5p easier in oligo and many things in Strand.

Need to decide how objects are going to finally be destroyed from the undoStack.  Right now, not sure if things will be left floating.