Added a virtualStrand accessor
findOverlappingRanges(self, qstrand, useCache=False)

to allow for finding complementary strands.  The method returns an generator that'll produce the complementary strands in order from lowest virtual index to highest virtual index.

Additionally, there is a caching system whereby the class stores an index based on the last index you left off at in this strand whereby there are bases in the the strand that exist outside the query strands range.  This is useful in the case of iterating left to right (in order) through all Virtual Scaffold strands and installing something on their complements.

Pretty sure it works, but certainly this needs testing