BIG bug fixes and feature completion, autobreak now works and autostaple generates oligos with loops with appropriate lengths:
Also wrote verifyOligo to verify an oligo's length matches the total length of all it's strands.  It's how I found the bugs to begin with in the following issues.

Added RemoveOligoCommand to remove Oligos because removing an Oligo with RemoveStrand causes the oligos to blow up and get added to the set of oligos in the part.  This appears to work.  We should use this in other instances probably and this will work great for oligo selections.

Also fixed CreateXoverCommand in when it increments length only when the xover being created doesn't create a loop.

Then fixed bugs in the processing of staplegraph output for the case of loops versus non loops.  Loops use the start index for the first split, and non-loops ignore this an go into the break items immediately.